requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.6.3
bcrypt==4.1.3
python-dotenv==1.0.0
APScheduler==3.10.4
Werkzeug==2.3.7
python-multipart==0.0.6
pycryptodome==3.19.0
PyJWT==2.10.1
pytest==7.4.3
pytest-mock==3.12.0
